New class: AsymptoticTestStatistic
Inherits from function. The main function is just a standard test
statistic. That is a function of (y, z, b) -> R^1. The `asymptotic` slot
is a backend function for the engine. e.g. xBalance is the backend for
the harmonic.mean.difference test statistic (this is implemented, though
requires more testing). Using the `type` = {"exact", "asymptotic"}
argument to the engine (perhaps via pRD), the user can select which of
the two she wants. The default is "exact." Obviously, for functions that
are no "AsymptoticTestStatistics", the only option is "exact".