Equipping xBalance to gracefully handle weird strata input (nlevels=0)
xBalance crashed a script I was running because it had been fed a
strata= argument one of the elements of which was a factor with no
levels.  (It had only NAs.)  In my situation as in any others like this
I could imagine, I'd have been happier to have it proceed with the
non-nonsensical strata.  So I've added exception-handling code to make
it do this. There is an accompanying addition to the tests file.