Fixing x86_64 linux/R 2.15.2 `R CMD check` bug
While it is not clear as to the reason, on Arch Linux x86_64 using R
2.15.2, the following would fail to find libraries installed in .local:

  R_LIBS=.local R --vanilla CMD check RItools.tar.gz

The R 2.15.2 release notes have some stuff about trying load packages
during the check process that are not on the usual library paths, which
may be related. Luckly add the `-l .local` flag to the check process
seesm to clear up the issue.