New class "AlignedCovs", refactored `alignDesignsByStrata`; "Design" |--> "DesignOptions"
I also added some explanation to the Rnw file re where this is going.
Stratum weights not currently expected to function if there are clusters
or non-uniform element weights (#65). The problem is that DesignWeights
isn't taking weights/element sizes into account.  Fixing that will require
passing this info down to DesignWeights, but without tying it to specific
variables and their NA patterns. At the moment, Design isn't set up to do
this, so I'm postponing for a later day.