Streamline DesignWeights calc, expectations
Clarified in internal documentation that stratum.weights argument should
give stratum weighting factors, not the full weights. (Stratum means of
unit weights get factored in by the DesignWeights calculation.)  Up to
this point, DesignWeights expected that if these were provided as raw
numbers then they'd be the complete weight, not just a weighting factor.

This permitted some streamlining of calculations internal to DesignWeights.
In the process I also updated internally facing docs for
`DesignWeights()` and inline comments for `designToDescriptives()`.