Streamlined designToDescriptives; in balanceTest, calculate stratum weights just once.
The streamlining removed the need for `balanceTest()` to invoke `DesignWeights()` twice,
once for descriptives calculations and then again separately for inferentials.  Perhaps
this will speed things up a bit.  This commit also eliminates a seemingly unnecessary
duplication of the object `aggDesigns` that's created within `balanceTest()`, perhaps
reducing memory usage in problems with large numbers of clusters.