Transparent, user-modifiable, once per invocation of balT() call to svd()
balTest() may evolve to offer the option of calculating the
same Mahalanobis (or reduced-rank Mahalanobis) imbalance summary
for each of multiple stratifications.  If so, it will be useful
to offer users a way to control the "meat matrix" used in the
Mahalanobis calculation, and it will also be helpful to expose
to the user what's being done.  The `X_rotation=` parameter
added to balT() in this commit represents one way to address
both of these aims. (Setting this up was the purpose of introducing
scale.DesignOptions a few commits ago.)

Note use of lazy evaluation, which may call for some further testing
and tweaking. I'm planning on immediately reverting this commit,
with intention of fishing it out again later, when I'm ready to make
use of it.