Expect post.alignment.transform to accept 2 args not 1
Also correct minor example boo-boo attributing
stratified test to Wilcoxon.  (Per my understanding,
the stratified version would be an aligned rank test.)
Anyway I took it out in this commit, as I needed to
add in a note about the 2 arg expectation for p.a.t.
and this was making the example comments rather busy.